refactor(fe/basic): split statement parsing
Moves statement parsing routines from Parser.cpp to Parser_Stmt.cpp to isolate program orchestration. Updates build scripts.

Pull/Commit Checklist:
- [x] Built on macOS/Linux (Debug).
- [x] ctest passed.
- [ ] Added/updated tests.
- [x] Code formatted; no new warnings.
- [x] No spec/API deviation (or ADR attached).
- [ ] Docs updated if behavior/API changed.
- [x] Updated docs/comments per policy.
- [x] Added/updated file headers.
